test: Enhance loop component tests with flow build and event handling utilities
